fix some warnings

+++ src/server/cachetable.h
@@ -101,7 +101,7 @@ private:
         return s & (_cols - 1);
     }
 public:
-    CacheTable(u32 slots, size_t cols)
+    CacheTable(u32 slots, u32 cols)
         : enabled(slots && cols), _cols(0), _rng(6581), _mask(0)
     {
         if(enabled)
